refactor(lib): make http feature optional
Split the default feature set to allow consumers to opt out of the http
dependency. The http feature is now explicitly required for http-related
functionality.

This change modifies the Cargo.toml to remove "http" from the default
features and updates the test module to conditionally import Error based
on the http feature.